moved interaction to region component
